+++ tortoise/contrib/mysql/fields.py
@@ -1,0 +1,5 @@
+from tortoise.fields import Field
+
+
+class GeometryField(Field):
+    SQL_TYPE = "GEOMETRY"