+++ CHANGELOG.rst
@@ -9,6 +9,11 @@ Changelog
 
 0.17
 ====
+0.17.2
+------
+- Add more `index` types.
+- Add `force_index`, `use_index` to `queryset`.
+
 0.17.1
 ------
 - Fix type for modules.